+++ packages/core/webpack/src/config/external.ts
@@ -5,13 +5,13 @@
 export const externalLibs = {
   "@floating-ui/react-dom-interactions": "@floating-ui/react-dom-interactions",
   "@nimbus-ds/styles": "@nimbus-ds/styles",
-  "@tiendanube/icons": "@tiendanube/icons",
   react: "react",
   "react-dom": "react-dom",
 };
 
 export const externalPackages = {
   "@nimbus-ds/icon": "@nimbus-ds/icon",
+  "@nimbus-ds/icons": "@nimbus-ds/icons",
   "@nimbus-ds/skeleton": "@nimbus-ds/skeleton",
   "@nimbus-ds/spinner": "@nimbus-ds/spinner",
   "@nimbus-ds/text": "@nimbus-ds/text",